+++ backend/main.py
@@ -3,11 +3,12 @@ from fastapi.middleware.cors import CORSMiddleware
 import config
 import models
 from database import engine
-from routers import manga, mal
+from routers import manga, mal, excelIO
 
 app = FastAPI()
 
 app.include_router(manga.router)
+app.include_router(excelIO.router)
 app.include_router(mal.router)
 
 # CORS settings
