+++ backend/routers/manga.py
@@ -139,133 +139,3 @@ def get_all_author_names(db: Session = Depends(get_db)) -> List[str]:
 @router.get("/authors/roles")
 def get_all_role_names(db: Session = Depends(get_db)) -> List[str]:
     return authorManager.get_all_role_names(db)
-
-
-############################################################################################################
-
-
-@router.get("/export")
-async def export_mangas_to_excel(db: Session = Depends(get_db)):
-    db_mangas = db.query(DBManga).all()
-    mangas = [mangaManager.create_manga_model(db, db_manga) for db_manga in db_mangas]
-
-    wb = Workbook()
-    ws = wb.active
-    ws.title = "Mangas"
-
-    # Add headers
-    ws.append(
-        [
-            "Title",
-            "Description",
-            "Genres",
-            "Authors(Roles)",
-            "Total Volumes",
-            "Specific Volumes",
-        ]
-    )
-
-    # Add data
-    for manga in mangas:
-        genres = ", ".join([genre.name for genre in manga.genres])
-        authors_roles = ", ".join(
-            [f"{author.name}({author.role})" for author in manga.authors]
-        )
-        specific_volumes = ", ".join(
-            [str(volume.volume_num) for volume in manga.volumes]
-        )
-
-        ws.append(
-            [
-                manga.title,
-                manga.description,
-                genres,
-                authors_roles,
-                len(manga.volumes),
-                specific_volumes,
-            ]
-        )
-
-    buffer = BytesIO()
-    wb.save(buffer)
-    buffer.seek(0)
-
-    return StreamingResponse(
-        buffer,
-        media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        headers={"Content-Disposition": "attachment; filename=mangas.xlsx"},
-    )
-
-
-@router.post("/import")
-async def import_mangas_from_excel(
-    file: UploadFile = File(...), db: Session = Depends(get_db)
-):
-    wb = load_workbook(filename=file.file)
-    ws = wb.active
-
-    for row in ws.iter_rows(min_row=2, values_only=True):
-        (
-            title,
-            description,
-            genres_str,
-            authors_roles_str,
-            total_volumes,
-            specific_volumes_str,
-        ) = row
-
-        manga_model = DBManga()
-        manga_model.title = title
-        manga_model.description = description
-        manga_model.totalVolumes = total_volumes
-        db.add(manga_model)
-        db.commit()
-        db.refresh(manga_model)
-
-        # Handle Genres
-        genre_names = (
-            [genre.strip() for genre in genres_str.split(",")] if genres_str else []
-        )
-        for genre_name in genre_names:
-            genre = genreManager.get_genre(db, genre_name)
-            if not genre:
-                genre = genreManager.create_genre(db, genre_name)
-            genreManager.create_relation(db, genre.id, manga_model.id)
-
-        authors_list = (
-            [author_role.strip() for author_role in authors_roles_str.split(",")]
-            if authors_roles_str
-            else []
-        )
-        #TODO: Watch out for cases like missing role etc, after hookup to frontend
-        # add error messages etc.
-        for author_role_str in authors_list:
-            parts = author_role_str.split("(")
-            author_name = parts[0].strip()
-            role = parts[1].replace(")", "").strip() if len(parts) > 1 else None
-            result_author = authorManager.get_author(db, author_name)
-            if result_author is None:
-                result_author = authorManager.create_author(db, author_name)
-            result_role = authorManager.get_role(db, role)
-            if result_role is None:
-                result_role = authorManager.create_role(db, role)
-            authorManager.create_relation(
-                db, result_author.id, manga_model.id, result_role.id
-            )
-
-        # Handle Volumes
-        volume_nums = (
-            [int(vol.strip()) for vol in specific_volumes_str.split(",")]
-            if specific_volumes_str
-            else []
-        )
-        for volume_num in volume_nums:
-            volume = Volume(
-                id=0,
-                volume_num=volume_num,
-                manga_id=manga_model.id,
-                cover_image="",
-            )
-            volumeManager.create_volume(db, volume)
-
-    return {"message": "Mangas imported successfully"}
